requirements.txt
# Django core
Django>=4.2

# WSGI & ASGI servers
gunicorn>=21.2.0
uvicorn[standard]>=0.23.0

# Security & TLS analysis
cryptography>=41.0.3
pyOpenSSL>=23.2.0

# ASN & IP reputation
ipwhois>=1.2.0
requests>=2.31.0

# JSON Schema validation
jsonschema>=4.19.0

# AI Behavior profiling
pandas>=2.0.3
scikit-learn>=1.3.0

# Database (Postgres)
psycopg2-binary>=2.9.6

# Utilities
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-extensions>=3.2.3
django-environ
django-widget-tweaks>=1.5.0
django-crontab

# Static files (optional, only if not using nginx)
whitenoise>=6.5.0

# Security (recommended)
django-cors-headers>=4.2.0
django-secure>=1.0.2

# Task queue (optional, for Celery/Redis if used)
redis>=5.0.0
celery>=5.3.1
